Struct weak_table::PtrWeakHashSet

pub struct PtrWeakHashSet<T, S = RandomState>(_);

A hash set with weak elements, hashed on element pointer.

When a weak pointer expires, its mapping is lazily removed.

Methods

impl<T: WeakElement> PtrWeakHashSet<T, RandomState> where
T::Strong: Deref,

pub fn new() -> Self

Creates an empty PtrWeakHashSet.

pub fn with_capacity(capacity: usize) -> Self

Creates an empty PtrWeakHashSet with the given capacity.

impl<T: WeakElement, S: BuildHasher> PtrWeakHashSet<T, S> where
T::Strong: Deref,

pub fn with_hasher(hash_builder: S) -> Self

Creates an empty PtrWeakHashSet with the given capacity and hasher.

pub fn with_capacity_and_hasher(capacity: usize, hash_builder: S) -> Self

Creates an empty PtrWeakHashSet with the given capacity and hasher.

pub fn hasher(&self) -> &S

Returns a reference to the map's BuildHasher.

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

pub fn remove_expired(&mut self)

Removes all mappings whose keys have expired.

pub fn reserve(&mut self, additional_capacity: usize)

Reserves room for additional elements.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity to the minimum allowed to hold the current number of elements.

pub fn len(&self) -> usize

Returns an over-approximation of the number of elements.

pub fn is_empty(&self) -> bool

Is the set known to be empty?

This could answer false for an empty set whose elements have expired but have yet to be collected.

pub fn load_factor(&self) -> f32

The proportion of buckets that are used.

This is an over-approximation because of expired elements.

pub fn clear(&mut self)

Removes all associations from the map.

pub fn contains(&self, key: &T::Strong) -> bool

Returns true if the map contains the specified key.

pub fn insert(&mut self, key: T::Strong) -> bool

Unconditionally inserts the value, returning the old value if already present. Does not replace the key.

pub fn remove(&mut self, key: &T::Strong) -> bool

Removes the entry with the given key, if it exists, and returns the value.

pub fn retain<F>(&mut self, f: F) where
F: FnMut(T::Strong) -> bool,

Removes all mappings not satisfying the given predicate.

Also removes any expired mappings.

pub fn is_subset<S1>(&self, other: &PtrWeakHashSet<T, S1>) -> bool where
S1: BuildHasher,

Is self a subset of other?

impl<T: WeakElement, S> PtrWeakHashSet<T, S> where
T::Strong: Deref,

Important traits for Iter<'a, T>
impl<'a, T: WeakElement> Iterator for Iter<'a, T> type Item = T::Strong;

pub fn iter(&self) -> Iter<T>

Gets an iterator over the keys and values.

Important traits for Drain<'a, T>
impl<'a, T: WeakElement> Iterator for Drain<'a, T> type Item = T::Strong;

pub fn drain(&mut self) -> Drain<T>

Gets a draining iterator, which removes all the values but retains the storage.