Struct weak_table::PtrWeakKeyHashMap

pub struct PtrWeakKeyHashMap<K, V, S = RandomState>(_);

A hash map with weak keys, hashed on key pointer.

When a weak pointer expires, its mapping is lazily removed.

Examples

use weak_table::PtrWeakKeyHashMap;
use std::rc::{Rc, Weak};

type Table = PtrWeakKeyHashMap<Weak<str>, usize>;

let mut map = Table::new();
let a = Rc::<str>::from("hello");
let b = Rc::<str>::from("hello");

map.insert(a.clone(), 5);

assert_eq!( map.get(&a), Some(&5) );
assert_eq!( map.get(&b), None );

map.insert(b.clone(), 7);

assert_eq!( map.get(&a), Some(&5) );
assert_eq!( map.get(&b), Some(&7) );

Implementations

impl<K: WeakElement, V> PtrWeakKeyHashMap<K, V, RandomState>where K::Strong: Deref,

pub fn new() -> Self

Creates an empty PtrWeakKeyHashMap.

O(1) time

pub fn with_capacity(capacity: usize) -> Self

Creates an empty PtrWeakKeyHashMap with the given capacity.

O(n) time

impl<K: WeakElement, V, S: BuildHasher> PtrWeakKeyHashMap<K, V, S>where K::Strong: Deref,

pub fn with_hasher(hash_builder: S) -> Self

Creates an empty PtrWeakKeyHashMap with the given capacity and hasher.

O(n) time

pub fn with_capacity_and_hasher(capacity: usize, hash_builder: S) -> Self

Creates an empty PtrWeakKeyHashMap with the given capacity and hasher.

O(n) time

pub fn hasher(&self) -> &S

Returns a reference to the map’s BuildHasher.

O(1) time

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

O(1) time

pub fn remove_expired(&mut self)

Removes all mappings whose keys have expired.

O(n) time

pub fn reserve(&mut self, additional_capacity: usize)

Reserves room for additional elements.

O(n) time

pub fn shrink_to_fit(&mut self)

Shrinks the capacity to the minimum allowed to hold the current number of elements.

O(n) time

pub fn len(&self) -> usize

Returns an over-approximation of the number of elements.

O(1) time

pub fn is_empty(&self) -> bool

Is the map known to be empty?

This could answer false for an empty map whose keys have expired but have yet to be collected.

O(1) time

pub fn load_factor(&self) -> f32

The proportion of buckets that are used.

This is an over-approximation because of expired keys.

O(1) time

pub fn entry(&mut self, key: K::Strong) -> Entry<'_, ByPtr<K>, V>

Gets the requested entry.

expected O(1) time; worst-case O(p) time

pub fn clear(&mut self)

Removes all associations from the map.

O(n) time

pub fn get(&self, key: &K::Strong) -> Option<&V>

Returns a reference to the value corresponding to the key.

expected O(1) time; worst-case O(p) time

pub fn contains_key(&self, key: &K::Strong) -> bool

Returns true if the map contains the specified key.

expected O(1) time; worst-case O(p) time

pub fn get_mut(&mut self, key: &K::Strong) -> Option<&mut V>

Returns a mutable reference to the value corresponding to the key.

expected O(1) time; worst-case O(p) time

pub fn insert(&mut self, key: K::Strong, value: V) -> Option<V>

Unconditionally inserts the value, returning the old value if already present. Does not replace the key.

expected O(1) time; worst-case O(p) time

pub fn remove(&mut self, key: &K::Strong) -> Option<V>

Removes the entry with the given key, if it exists, and returns the value.

expected O(1) time; worst-case O(p) time

pub fn retain<F>(&mut self, f: F)where F: FnMut(K::Strong, &mut V) -> bool,

Removes all mappings not satisfying the given predicate.

Also removes any expired mappings.

O(n) time

pub fn submap_with<F, S1, V1>( &self, other: &PtrWeakKeyHashMap<K, V1, S1>, value_equal: F ) -> boolwhere F: FnMut(&V, &V1) -> bool, S1: BuildHasher,

Is this map a submap of the other, using the given value comparison.

In particular, all the keys of self must be in other and the values must compare true with value_equal.

expected O(n) time; worst-case O(nq) time (where n is self.capacity() and q is the length of the probe sequences in other)

pub fn is_submap<V1, S1>(&self, other: &PtrWeakKeyHashMap<K, V1, S1>) -> boolwhere V: PartialEq<V1>, S1: BuildHasher,

Is self a submap of other?

expected O(n) time; worst-case O(nq) time (where n is self.capacity() and q is the length of the probe sequences in other)

pub fn domain_is_subset<V1, S1>( &self, other: &PtrWeakKeyHashMap<K, V1, S1> ) -> boolwhere S1: BuildHasher,

Are the keys of self a subset of the keys of other?

expected O(n) time; worst-case O(nq) time (where n is self.capacity() and q is the length of the probe sequences in other)

impl<K: WeakElement, V, S> PtrWeakKeyHashMap<K, V, S>where K::Strong: Deref,

pub fn iter(&self) -> Iter<'_, ByPtr<K>, V>

Gets an iterator over the keys and values.

O(1) time

pub fn keys(&self) -> Keys<'_, ByPtr<K>, V>

Gets an iterator over the keys.

O(1) time

pub fn values(&self) -> Values<'_, ByPtr<K>, V>

Gets an iterator over the values.

O(1) time

pub fn iter_mut(&mut self) -> IterMut<'_, ByPtr<K>, V>

Gets an iterator over the keys and mutable values.

O(1) time

pub fn values_mut(&mut self) -> ValuesMut<'_, ByPtr<K>, V>

Gets an iterator over the mutable values.

O(1) time

pub fn drain(&mut self) -> Drain<'_, ByPtr<K>, V>

Gets a draining iterator, which removes all the values but retains the storage.

O(1) time (and O(n) time to dispose of the result)