Struct weak_table::PtrWeakWeakHashMap

pub struct PtrWeakWeakHashMap<K, V, S = RandomState>(_);

A hash map with weak keys and weak values, hashed on key pointer.

When a weak pointer expires, its mapping is lazily removed.

Methods

impl<K: WeakElement, V: WeakElement> PtrWeakWeakHashMap<K, V, RandomState> where
K::Strong: Deref,

pub fn new() -> Self

Creates an empty PtrWeakWeakHashMap.

pub fn with_capacity(capacity: usize) -> Self

Creates an empty PtrWeakWeakHashMap with the given capacity.

impl<K: WeakElement, V: WeakElement, S: BuildHasher> PtrWeakWeakHashMap<K, V, S> where
K::Strong: Deref,

pub fn with_hasher(hash_builder: S) -> Self

Creates an empty PtrWeakWeakHashMap with the given capacity and hasher.

pub fn with_capacity_and_hasher(capacity: usize, hash_builder: S) -> Self

Creates an empty PtrWeakWeakHashMap with the given capacity and hasher.

pub fn hasher(&self) -> &S

Returns a reference to the map's BuildHasher.

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

pub fn remove_expired(&mut self)

Removes all mappings whose keys have expired.

pub fn reserve(&mut self, additional_capacity: usize)

Reserves room for additional elements.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity to the minimum allowed to hold the current number of elements.

pub fn len(&self) -> usize

Returns an over-approximation of the number of elements.

pub fn load_factor(&self) -> f32

The proportion of buckets that are used.

This is an over-approximation because of expired keys.

pub fn entry(&mut self, key: K::Strong) -> Entry<ByPtr<K>, V>

Gets the requested entry.

pub fn clear(&mut self)

Removes all associations from the map.

pub fn get(&self, key: &K::Strong) -> Option<V::Strong>

Returns a reference to the value corresponding to the key.

pub fn contains_key(&self, key: &K::Strong) -> bool

Returns true if the map contains the specified key.

pub fn insert(&mut self, key: K::Strong, value: V::Strong) -> Option<V::Strong>

Unconditionally inserts the value, returning the old value if already present. Does not replace the key.

pub fn remove(&mut self, key: &K::Strong) -> Option<V::Strong>

Removes the entry with the given key, if it exists, and returns the value.

pub fn retain<F>(&mut self, f: F) where
F: FnMut(K::Strong, V::Strong) -> bool,

Removes all mappings not satisfying the given predicate.

Also removes any expired mappings.

pub fn submap_with<F, S1, V1>(
&self,
other: &PtrWeakWeakHashMap<K, V1, S1>,
value_equal: F
) -> bool where
F: FnMut(V::Strong, V1::Strong) -> bool,
V1: WeakElement,
S1: BuildHasher,

Is this map a submap of the other, using the given value comparison.

In particular, all the keys of self must be in other and the values must compare true with value_equal.

pub fn is_submap<V1, S1>(&self, other: &PtrWeakWeakHashMap<K, V1, S1>) -> bool where
V1: WeakElement,
V::Strong: PartialEq<V1::Strong>,
S1: BuildHasher,

Is self a submap of other?

pub fn domain_is_subset<V1, S1>(
&self,
other: &PtrWeakWeakHashMap<K, V1, S1>
) -> bool where
V1: WeakElement,
S1: BuildHasher,

Are the keys of self a subset of the keys of other?

impl<K: WeakElement, V: WeakElement, S> PtrWeakWeakHashMap<K, V, S> where
K::Strong: Deref,

Important traits for Iter<'a, K, V>
impl<'a, K: WeakElement, V: WeakElement> Iterator for Iter<'a, K, V> type Item = (K::Strong, V::Strong);

pub fn iter(&self) -> Iter<ByPtr<K>, V>

Gets an iterator over the keys and values.

Important traits for Keys<'a, K, V>
impl<'a, K: WeakElement, V: WeakElement> Iterator for Keys<'a, K, V> type Item = K::Strong;

pub fn keys(&self) -> Keys<ByPtr<K>, V>

Gets an iterator over the keys.

Important traits for Values<'a, K, V>
impl<'a, K: WeakElement, V: WeakElement> Iterator for Values<'a, K, V> type Item = V::Strong;

pub fn values(&self) -> Values<ByPtr<K>, V>

Gets an iterator over the values.

Important traits for Drain<'a, K, V>
impl<'a, K: WeakElement, V: WeakElement> Iterator for Drain<'a, K, V> type Item = (K::Strong, V::Strong);

pub fn drain(&mut self) -> Drain<ByPtr<K>, V>

Gets a draining iterator, which removes all the values but retains the storage.