Struct weak_table::WeakKeyHashMap

pub struct WeakKeyHashMap<K, V, S = RandomState> { /* fields omitted */ }

A hash map with weak keys, hashed on key value.

When a weak pointer expires, its mapping is lazily removed.

Methods

impl<K: WeakKey, V> WeakKeyHashMap<K, V, RandomState>

pub fn new() -> Self

Creates an empty WeakKeyHashMap.

pub fn with_capacity(capacity: usize) -> Self

Creates an empty WeakKeyHashMap with the given capacity.

impl<K: WeakKey, V, S: BuildHasher> WeakKeyHashMap<K, V, S>

pub fn with_hasher(hash_builder: S) -> Self

Creates an empty WeakKeyHashMap with the given capacity and hasher.

pub fn with_capacity_and_hasher(capacity: usize, hash_builder: S) -> Self

Creates an empty WeakKeyHashMap with the given capacity and hasher.

pub fn hasher(&self) -> &S

Returns a reference to the map's BuildHasher.

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

pub fn remove_expired(&mut self)

Removes all mappings whose keys have expired.

pub fn reserve(&mut self, additional_capacity: usize)

Reserves room for additional elements.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity to the minimum allowed to hold the current number of elements.

pub fn len(&self) -> usize

Returns an over-approximation of the number of elements.

pub fn is_empty(&self) -> bool

Is the map empty?

Note that this may return false even if all keys in the map have expired, if they haven't been collected yet.

pub fn load_factor(&self) -> f32

The proportion of buckets that are used.

This is an over-approximation because of expired keys.

pub fn entry(&mut self, key: K::Strong) -> Entry<K, V>

Gets the requested entry.

pub fn clear(&mut self)

Removes all associations from the map.

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<&V> where
Q: Hash + Eq,
K::Key: Borrow<Q>,

Returns a reference to the value corresponding to the key.

pub fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool where
Q: Hash + Eq,
K::Key: Borrow<Q>,

Returns true if the map contains the specified key.

pub fn get_key<Q: ?Sized>(&self, key: &Q) -> Option<K::Strong> where
Q: Hash + Eq,
K::Key: Borrow<Q>,

Returns a strong reference to the key, if found.

pub fn get_both<Q: ?Sized>(&self, key: &Q) -> Option<(K::Strong, &V)> where
Q: Hash + Eq,
K::Key: Borrow<Q>,

Returns a pair of a strong reference to the key, and a reference to the value, if present.

pub fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut V> where
Q: Hash + Eq,
K::Key: Borrow<Q>,

Returns a mutable reference to the value corresponding to the key.

pub fn get_both_mut<Q: ?Sized>(
&mut self,
key: &Q
) -> Option<(K::Strong, &mut V)> where
Q: Hash + Eq,
K::Key: Borrow<Q>,

Returns a pair of a strong reference to the key, and a mutable reference to the value, if present.

pub fn insert(&mut self, key: K::Strong, value: V) -> Option<V>

Unconditionally inserts the value, returning the old value if already present.

Unlike std::collections::HashMap, this replaced the key even if occupied.

pub fn remove<Q: ?Sized>(&mut self, key: &Q) -> Option<V> where
Q: Hash + Eq,
K::Key: Borrow<Q>,

Removes the entry with the given key, if it exists, and returns the value.

pub fn retain<F>(&mut self, f: F) where
F: FnMut(K::Strong, &mut V) -> bool,

Removes all mappings not satisfying the given predicate.

Also removes any expired mappings.

pub fn is_submap_with<F, S1, V1>(
&self,
other: &WeakKeyHashMap<K, V1, S1>,
value_equal: F
) -> bool where
F: FnMut(&V, &V1) -> bool,
S1: BuildHasher,

Is this map a submap of the other, using the given value comparison.

In particular, all the keys of self must be in other and the values must compare true with value_equal.

pub fn is_submap<V1, S1>(&self, other: &WeakKeyHashMap<K, V1, S1>) -> bool where
V: PartialEq<V1>,
S1: BuildHasher,

Is self a submap of other?

pub fn domain_is_subset<V1, S1>(
&self,
other: &WeakKeyHashMap<K, V1, S1>
) -> bool where
S1: BuildHasher,

Are the keys of self a subset of the keys of other?

impl<K: WeakElement, V, S> WeakKeyHashMap<K, V, S>

Important traits for Iter<'a, K, V>
impl<'a, K: WeakElement, V> Iterator for Iter<'a, K, V> type Item = (K::Strong, &'a V);

pub fn iter(&self) -> Iter<K, V>

Gets an iterator over the keys and values.

Important traits for Keys<'a, K, V>
impl<'a, K: WeakElement, V> Iterator for Keys<'a, K, V> type Item = K::Strong;

pub fn keys(&self) -> Keys<K, V>

Gets an iterator over the keys.

Important traits for Values<'a, K, V>
impl<'a, K: WeakElement, V> Iterator for Values<'a, K, V> type Item = &'a V;

pub fn values(&self) -> Values<K, V>

Gets an iterator over the values.

Important traits for IterMut<'a, K, V>
impl<'a, K: WeakElement, V> Iterator for IterMut<'a, K, V> type Item = (K::Strong, &'a mut V);

pub fn iter_mut(&mut self) -> IterMut<K, V>

Gets an iterator over the keys and mutable values.

Important traits for ValuesMut<'a, K, V>
impl<'a, K: WeakElement, V> Iterator for ValuesMut<'a, K, V> type Item = &'a mut V;

pub fn values_mut(&mut self) -> ValuesMut<K, V>

Gets an iterator over the mutable values.

Important traits for Drain<'a, K, V>
impl<'a, K: WeakElement, V> Iterator for Drain<'a, K, V> type Item = (K::Strong, V);

pub fn drain(&mut self) -> Drain<K, V>

Gets a draining iterator, which removes all the values but retains the storage.