Struct weak_table::WeakValueHashMap

pub struct WeakValueHashMap<K, V, S = RandomState> { /* fields omitted */ }

A hash map with weak values.

When a weak pointer expires, its mapping is lazily removed.

Methods

impl<K: Eq + Hash, V: WeakElement> WeakValueHashMap<K, V, RandomState>

pub fn new() -> Self

Creates an empty WeakValueHashMap.

pub fn with_capacity(capacity: usize) -> Self

Creates an empty WeakValueHashMap with the given capacity.

impl<K: Eq + Hash, V: WeakElement, S: BuildHasher> WeakValueHashMap<K, V, S>

pub fn with_hasher(hash_builder: S) -> Self

Creates an empty WeakValueHashMap with the given capacity and hasher.

pub fn with_capacity_and_hasher(capacity: usize, hash_builder: S) -> Self

Creates an empty WeakValueHashMap with the given capacity and hasher.

pub fn hasher(&self) -> &S

Returns a reference to the map's BuildHasher.

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

pub fn remove_expired(&mut self)

Removes all mappings whose keys have expired.

pub fn reserve(&mut self, additional_capacity: usize)

Reserves room for additional elements.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity to the minimum allowed to hold the current number of elements.

pub fn len(&self) -> usize

Returns an over-approximation of the number of elements.

pub fn is_empty(&self) -> bool

Is the map empty?

Note that this may return false even if all keys in the map have expired, if they haven't been collected yet.

pub fn load_factor(&self) -> f32

The proportion of buckets that are used.

This is an over-approximation because of expired keys.

pub fn entry(&mut self, key: K) -> Entry<K, V>

Gets the requested entry.

pub fn clear(&mut self)

Removes all associations from the map.

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<V::Strong> where
Q: Hash + Eq,
K: Borrow<Q>,

Returns a reference to the value corresponding to the key.

pub fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool where
Q: Hash + Eq,
K: Borrow<Q>,

Returns true if the map contains the specified key.

pub fn insert(&mut self, key: K, value: V::Strong) -> Option<V::Strong>

Unconditionally inserts the value, returning the old value if already present.

Like std::collections::HashMap, this does not replace the key if occupied.

pub fn remove<Q: ?Sized>(&mut self, key: &Q) -> Option<V::Strong> where
Q: Hash + Eq,
K: Borrow<Q>,

Removes the entry with the given key, if it exists, and returns the value.

pub fn retain<F>(&mut self, f: F) where
F: FnMut(&K, V::Strong) -> bool,

Removes all mappings not satisfying the given predicate.

Also removes any expired mappings.

pub fn is_submap_with<F, S1, V1>(
&self,
other: &WeakValueHashMap<K, V1, S1>,
value_equal: F
) -> bool where
V1: WeakElement,
F: FnMut(V::Strong, V1::Strong) -> bool,
S1: BuildHasher,

Is this map a submap of the other, using the given value comparison.

In particular, all the keys of self must be in other and the values must compare true with value_equal.

pub fn is_submap<V1, S1>(&self, other: &WeakValueHashMap<K, V1, S1>) -> bool where
V1: WeakElement,
V::Strong: PartialEq<V1::Strong>,
S1: BuildHasher,

Is self a submap of other?

pub fn domain_is_subset<V1, S1>(
&self,
other: &WeakValueHashMap<K, V1, S1>
) -> bool where
V1: WeakElement,
S1: BuildHasher,

Are the keys of self a subset of the keys of other?

impl<K, V: WeakElement, S> WeakValueHashMap<K, V, S>

Important traits for Iter<'a, K, V>
impl<'a, K, V: WeakElement> Iterator for Iter<'a, K, V> type Item = (&'a K, V::Strong);

pub fn iter(&self) -> Iter<K, V>

Gets an iterator over the keys and values.

Important traits for Keys<'a, K, V>
impl<'a, K, V: WeakElement> Iterator for Keys<'a, K, V> type Item = &'a K;

pub fn keys(&self) -> Keys<K, V>

Gets an iterator over the keys.

Important traits for Values<'a, K, V>
impl<'a, K, V: WeakElement> Iterator for Values<'a, K, V> type Item = V::Strong;

pub fn values(&self) -> Values<K, V>

Gets an iterator over the values.

Important traits for Drain<'a, K, V>
impl<'a, K, V: WeakElement> Iterator for Drain<'a, K, V> type Item = (K, V::Strong);

pub fn drain(&mut self) -> Drain<K, V>

Gets a draining iterator, which removes all the values but retains the storage.