Module weak_table::weak_key_hash_map

A hash map where the keys are held by weak pointers and compared by key value.

Re-exports

pub use super::WeakKeyHashMap;

Structs

Drain
An iterator that consumes the values of a weak hash map, leaving it empty.

IntoIter
An iterator that consumes the values of a weak hash map, leaving it empty.

Iter
An iterator over the keys and values of the weak hash map.

IterMut
An iterator over the keys and mutable values of the weak hash map.

Keys
An iterator over the keys of the weak hash map.

OccupiedEntry
An occupied entry, which can be removed or viewed.

VacantEntry
A vacant entry, which can be inserted in or viewed.

Values
An iterator over the values of the weak hash map.

ValuesMut
An iterator over the mutable values of the weak hash map.

Enums

Entry
Represents an entry in the table which may be occupied or vacant.