Struct weak_table::weak_weak_hash_map::OccupiedEntry

pub struct OccupiedEntry<'a, K: 'a + WeakKey, V: 'a + WeakElement> { /* fields omitted */ }

An occupied entry, which can be removed or viewed.

Methods

impl<'a, K: WeakKey, V: WeakElement> OccupiedEntry<'a, K, V>

pub fn key(&self) -> &K::Strong

Gets a reference to the key held by the entry.

pub fn remove_entry(self) -> (K::Strong, V::Strong)

Takes ownership of the key and value from the map.

pub fn get(&self) -> &V::Strong

Gets a reference to the value in the entry.

pub fn get_strong(&self) -> V::Strong

Gets a clone of the reference to the value in the entry.

pub fn insert(&mut self, value: V::Strong) -> V::Strong

Replaces the value in the entry with the given value.

pub fn remove(self) -> V::Strong

Removes the entry, returning the value.